feat: ledger support; refactoring
